use a session cache to mimise startup times
Each time a session is created and used for the first time there is massive overhead
leading to execution times of O(5s) for images. But if the same session is reused then
this drops to execution times O(50ms) on each subsequent run after the session has been
used.